+++ examples/using_arcs.py
@@ -4,14 +4,36 @@
 # License: MIT License
 from __future__ import unicode_literals
 import ezdxf
+from ezdxf.algebra import Vector, Arc, UCS
 
 dwg = ezdxf.new('R2000')
 modelspace = dwg.modelspace()
 
+# create a 2D arcs in xy-plane
 delta = 30
 for count in range(12):
     modelspace.add_arc(center=(0, 0), radius=10+count, start_angle=count*delta, end_angle=(count+1)*delta)
 
+# create a 3D arc from 3 points in WCS
+start_point_wcs = Vector(3, 0, 0)
+end_point_wcs = Vector(0, 3, 0)
+def_point_wcs = Vector(0, 0, 3)
+
+# create UCS
+ucs = UCS.from_x_axis_and_point_in_xy(origin=def_point_wcs, axis=start_point_wcs-def_point_wcs, point=end_point_wcs)
+start_point_ucs = ucs.from_wcs(start_point_wcs)
+end_point_ucs = ucs.from_wcs(end_point_wcs)
+def_point_ucs = Vector(0, 0)  # origin of UCS
+
+# create arc in the xy-plane of the UCS
+arc = Arc.from_3p(start_point_ucs, end_point_ucs, def_point_ucs)
+arc.add_to_layout(modelspace, ucs, dxfattribs={'color': 1})  # red arc
+
+arc = Arc.from_3p(end_point_ucs, start_point_ucs, def_point_ucs)
+arc.add_to_layout(modelspace, ucs, dxfattribs={'color': 2})  # yellow arc
+
+
+# saving DXF file
 filename = 'using_arcs.dxf'
 dwg.saveas(filename)
 print("drawing '%s' created.\n" % filename)