+++ NEWS.rst
@@ -26,7 +26,8 @@ Version 0.8.9 - dev
     - NEW: Layout.set_redraw_order() and Layout.get_redraw_order(), to change redraw order of entities in model space and
       paper space layouts
     - NEW: BlockLayout.is_layout_block, True if block is a model space or paper space block definition
-    - NEW: ezdxf.algebra.Arc helper class to create arcs from 2 points and an angle or from 3 points
+    - NEW: ezdxf.algebra.Arc helper class to create arcs from 2 points and an angle or radius, or from 3 points
+    - NEW: ezdxf.algebra.Arc.add_to_layout() with UCS support to create 3D arcs
     - CHANGE: Drawing.blocks.delete_block(name, safe=True), new parameter save, check if block is still referenced
       (raises DXFValueError)
     - CHANGE: Drawing.blocks.delete_all_blocks(safe=True), new parameter save, ignores blocks still referenced if safe is True
@@ -49,7 +50,7 @@ Version 0.8.9 - dev
         - DICTIONARYVAR
         - DIMASSOC
         - FIELD (incomplete)
-        - FIELDLIST (not documented by Autodesk, with packed data)
+        - FIELDLIST (not documented by Autodesk)
         - IDBUFFER
         - LAYER_FILTER
         - MATERIAL